Service layer that handles returing IResult for all job ad repository methods. Enabling dependency injection, testing and scalability.
One method for each of the methods. Tries to get a list of JobDto (SearchJob) or a specific job from id (GetJobFromId) and returns the mapped JSON back to frontend or a bad request with error message.

Last method (GetJobFromIdChatGpt) is only to be used internally in the program for out letter generator. It is different from the other get from id as it returns another DTO with less info and not an IResult. Otherwise does the same things.