Renamed variables, changed and moved exceptions.
Renamed variables for clarification and readability:  jobId-> adId (as it's the ads id we're using not a job).
offset -> pageIndex (just felt easier to understand)

Also removed pageIndex nullcheck as that is now done directly in the query.

Changed InvalidOperation to Json exception as it makes more sense.

Also moved the try catches out of the repository and into the service as it felt unecessary to return empty lists when an exception occured in here. Much more sense to just throw exceptions in the repo and catch them above in the service.