Added endpoints for email verification, changed identity settings for login, and removed unnused usings.
-Added endpoints for resetting forgotten password. More detailed info on those in the UserRepository commit where the business logic is.
Also commented away the change password endpoint to avoid confusion as it does the same as reset password but for a already logged in user and without email verification. Unsure if this is something we need right now.

-Changed identity settings so we now check if email is confirmed and not account as that's all we're verify at the moment.
As well as adding lockout setting so user is prohibited from logging in for 10 mins after 8 failed attempts.