Added methods for email verification and forgot- and reset password, changed login and acc creation logic ,removed unecessary usings, commented away change password.
-For login we changed to using SignInManager instead of UserManager to enable the morw advanced security like automatic checks of email verification and lockout etc. Allows for future scalability ad security.

-Continuing with login we now lock the account according to the identity settings in main (currently 10mins after 8 failed logins).

Thanks to SingInManager and email now must be confirmed to login (this was an error it did not check before this change).

Also added some more specific exceptions if login failed thanks to this change.

-In CreateAccount I changed faulty login credentials to ArgumentExceptions as that makes more sense according to the definiton.

Added "Uri.EscapeDataString" to the confirmation token in the callbackUrl. This is so none of the special chars may interfere with the link as without this they can be inturperated as other things in the link.

Changed the email subject and body to be more in detail but is still a scrambled mess, probably needs to be html here but we're probably gonna fix this futher down the line anyway. This is just to test.

Fixed issue where the method always returned true if no errors where thrown. Now EmailSender actually contains logic to send back if the process was sucessful or not.

-Added method to verify email from the token that is generated upon account creation.

-Added password reset token generation method that works similarly to the account creation with the email. Creates a token and sends it in an email to the inputted email if there is a user with that email. Code is then verified in the ResetPassword method.

-Added reset password method. Takes the password reset code and checks if it is valid then changes the password to the one sent in.

-Commented away change password until we know if we're gonna use it.