Added the logic to all the get/save/delete cover letter methods
Self explanatory but created the method to save cover letters. Retrieves the user, then checks if cover letter we're trying to save already exists. If yes we overwrite it as to update it for potential regeneration.
Then just saves it. Using a db transaction to rollback any changes in case it fails.

Likewise with the remove method, retrieves the user and the cover letter.
Then attempt to remove it from the user and the context. Rollback if fail.

Get method just retrieves the existing data if any.